+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/parameters/Parameter.kt
@@ -197,7 +197,7 @@ internal fun List<IrValueParameter>.mapToConstructorParameters(
   remapper: TypeRemapper = NOOP_TYPE_REMAPPER
 ): List<Parameter> {
   return map { valueParameter ->
-    valueParameter.toConstructorParameter(IrParameterKind.Regular, remapper)
+    valueParameter.toConstructorParameter(valueParameter.kind, remapper)
   }
 }
 
@@ -210,8 +210,13 @@ internal fun IrValueParameter.toConstructorParameter(
   // type mangling
   val declaredType = remapper.remapType(this@toConstructorParameter.type)
 
-  val contextKey = declaredType.asContextualTypeKey(qualifierAnnotation(),
-    hasMetroDefault(), patchMutableCollections = false, declaration = this)
+  val contextKey =
+    declaredType.asContextualTypeKey(
+      qualifierAnnotation(),
+      hasMetroDefault(),
+      patchMutableCollections = false,
+      declaration = this,
+    )
 
   val assistedAnnotation = annotationsIn(context.metroSymbols.assistedAnnotations).singleOrNull()
 
@@ -285,7 +290,12 @@ internal fun IrProperty.toMemberInjectParameter(
       getter?.body ?: backingField?.initializer
     }
   val contextKey =
-    declaredType.asContextualTypeKey(with(context) { qualifierAnnotation() }, defaultValue != null, patchMutableCollections = false, declaration = this)
+    declaredType.asContextualTypeKey(
+      with(context) { qualifierAnnotation() },
+      defaultValue != null,
+      patchMutableCollections = false,
+      declaration = this,
+    )
 
   return Parameter.member(
     kind = kind,
@@ -309,7 +319,12 @@ internal fun IrValueParameter.toMemberInjectParameter(
       ?: this@toMemberInjectParameter.type
 
   val contextKey =
-    declaredType.asContextualTypeKey(with(context) { qualifierAnnotation() }, defaultValue != null, patchMutableCollections = false, declaration = this)
+    declaredType.asContextualTypeKey(
+      with(context) { qualifierAnnotation() },
+      defaultValue != null,
+      patchMutableCollections = false,
+      declaration = this,
+    )
 
   return Parameter.member(
     kind = kind,