+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/InjectConstructorTransformer.kt
@@ -10,6 +10,7 @@ import dev.zacsweers.metro.compiler.generatedClass
 import dev.zacsweers.metro.compiler.ir.ClassFactory
 import dev.zacsweers.metro.compiler.ir.IrMetroContext
 import dev.zacsweers.metro.compiler.ir.assignConstructorParamsToFields
+import dev.zacsweers.metro.compiler.ir.contextParameters
 import dev.zacsweers.metro.compiler.ir.createIrBuilder
 import dev.zacsweers.metro.compiler.ir.dispatchReceiverFor
 import dev.zacsweers.metro.compiler.ir.finalizeFakeOverride
@@ -376,17 +377,68 @@ internal class InjectConstructorTransformer(
             .symbol
       }
 
-      // TODO
-      //  copy default values
       invokeFunction.apply {
         val functionReceiver = dispatchReceiverParameter!!
         body =
           pluginContext.createIrBuilder(symbol).run {
+            val sourceParameters = targetCallable.owner.parameters()
+            if (invokeFunction.origin == Origins.TopLevelInjectFunctionClassFunction) {
+              // If this is a top-level function, we need to patch up the parameters
+              // TODO this crashes kotlinc when it generates $default functions :(
+              //  https://youtrack.jetbrains.com/issue/KT-81656/
+              //  copyParameterDefaultValues(
+              //    providerFunction = null,
+              //    sourceMetroParameters = sourceParameters,
+              //    sourceParameters = sourceParameters.nonDispatchParameters.filter {
+              //     it.isAssisted }.map { it.ir },
+              //    targetParameters = invokeFunction.nonDispatchParameters,
+              //    targetGraphParameter = null,
+              //    wrapInProvider = false,
+              //    isTopLevelFunction = true,
+              //  )
+            }
+
             val constructorParameterNames =
               constructorParameters.regularParameters.associateBy { it.originalName }
 
+            val contextParameterNames =
+              invokeFunction.contextParameters.associate { it.name to irGet(it) }
+
             val functionParamsByName =
               invokeFunction.regularParameters.associate { it.name to irGet(it) }
+
+            val contextArgs =
+              sourceParameters.contextParameters.map { targetParam ->
+                when (val parameterName = targetParam.originalName) {
+                  in constructorParameterNames -> {
+                    val constructorParam = constructorParameterNames.getValue(parameterName)
+                    val providerInstance =
+                      irGetField(
+                        irGet(functionReceiver),
+                        constructorParametersToFields.getValue(constructorParam),
+                      )
+                    val contextKey = targetParam.contextualTypeKey
+                    typeAsProviderArgument(
+                      contextKey = contextKey,
+                      bindingCode = providerInstance,
+                      isAssisted = false,
+                      isGraphInstance = constructorParam.isGraphInstance,
+                    )
+                  }
+
+                  in functionParamsByName -> {
+                    functionParamsByName.getValue(targetParam.originalName)
+                  }
+
+                  in contextParameterNames -> {
+                    contextParameterNames.getValue(targetParam.originalName)
+                  }
+
+                  else -> {
+                    error("Unmatched top level injected function param: $targetParam")
+                  }
+                }
+              }
 
             val args =
               targetCallable.owner.parameters().regularParameters.map { targetParam ->
@@ -423,6 +475,7 @@ internal class InjectConstructorTransformer(
                 extensionReceiver = null,
                 typeHint = targetCallable.owner.returnType,
                 // TODO type params
+                contextArgs = contextArgs,
                 args = args,
               )
 
@@ -469,6 +522,7 @@ internal class InjectConstructorTransformer(
     val newInstanceFunction =
       generateStaticNewInstanceFunction(
         parentClass = classToGenerateCreatorsIn,
+        sourceMetroParameters = constructorParameters,
         sourceParameters = constructorParameters.regularParameters.map { it.ir },
       ) { function ->
         irCallConstructor(