+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/factoryIrUtil.kt
@@ -41,6 +41,7 @@ import org.jetbrains.kotlin.ir.util.copyTypeParametersFrom
 import org.jetbrains.kotlin.ir.util.deepCopyWithSymbols
 import org.jetbrains.kotlin.ir.util.functions
 import org.jetbrains.kotlin.ir.util.isObject
+import org.jetbrains.kotlin.ir.util.nonDispatchParameters
 import org.jetbrains.kotlin.ir.util.parentAsClass
 
 /**
@@ -68,10 +69,11 @@ internal fun generateStaticCreateFunction(
   return function.apply {
     if (patchCreationParams) {
       val instanceParam = regularParameters.find { it.origin == Origins.InstanceParameter }
-      val valueParamsToPatch = regularParameters.filter { it.origin == Origins.RegularParameter }
+      val valueParamsToPatch = nonDispatchParameters.filter { it.origin == Origins.RegularParameter }
       copyParameterDefaultValues(
         providerFunction = providerFunction,
-        sourceParameters = parameters.regularParameters.filterNot { it.isAssisted }.map { it.ir },
+        sourceMetroParameters = parameters,
+        sourceParameters = parameters.nonDispatchParameters.filterNot { it.isAssisted }.map { it.ir },
         targetParameters = valueParamsToPatch,
         targetGraphParameter = instanceParam,
         wrapInProvider = true,
@@ -109,6 +111,7 @@ internal fun generateStaticCreateFunction(
 context(context: IrMetroContext)
 internal fun generateStaticNewInstanceFunction(
   parentClass: IrClass,
+  sourceMetroParameters: Parameters,
   sourceParameters: List<IrValueParameter>,
   targetFunction: IrFunction? = null,
   buildBody: IrBuilderWithScope.(IrSimpleFunction) -> IrExpression,
@@ -117,9 +120,10 @@ internal fun generateStaticNewInstanceFunction(
 
   return function.apply {
     val instanceParam = regularParameters.find { it.origin == Origins.InstanceParameter }
-    val valueParametersToMap = regularParameters.filter { it.origin == Origins.RegularParameter }
+    val valueParametersToMap = nonDispatchParameters.filter { it.origin == Origins.RegularParameter }
     copyParameterDefaultValues(
       providerFunction = targetFunction,
+      sourceMetroParameters = sourceMetroParameters,
       sourceParameters = sourceParameters,
       targetParameters = valueParametersToMap,
       targetGraphParameter = instanceParam,